Move Analytics option filter tests to new SettingsTest.
